Use naked attribute to implement naked asm
