+++ dmd2/declaration.c
@@ -795,8 +795,6 @@ VarDeclaration::VarDeclaration(Loc loc, Type *type, Identifier *id, Initializer 
 #if IN_LLVM
     aggrIndex = 0;
 
-    nakedUse = false;
-
     availableExternally = true; // assume this unless proven otherwise
 #endif
 }
