+++ gen/functions.cpp
@@ -777,7 +777,19 @@ void DtoDeclareFunction(FuncDeclaration* fdecl)
             func->addFnAttr(NoRedZone);
 #endif
         }
+
+        if (fdecl->naked)
+        {
+#if LDC_LLVM_VER >= 303
+            func->addFnAttr(llvm::Attribute::Naked);
+#elif LDC_LLVM_VER == 302
+            func->addFnAttr(llvm::Attributes::Naked);
+#else
+            func->addFnAttr(Naked);
+#endif
+        }
     }
+
 
     // main
     if (fdecl->isMain()) {
@@ -932,13 +944,6 @@ void DtoDefineFunction(FuncDeclaration* fd)
     if (Logger::enabled())
         Logger::println("DtoDefineFunc(%s): %s", fd->toPrettyChars(), fd->loc.toChars());
     LOG_SCOPE;
-
-    // if this function is naked, we take over right away! no standard processing!
-    if (fd->naked)
-    {
-        DtoDefineNakedFunction(fd);
-        return;
-    }
 
     // debug info
     fd->ir.irFunc->diSubprogram = gIR->DBuilder.EmitSubProgram(fd);
@@ -1240,4 +1245,159 @@ bool FuncDeclaration::isVaIntrinsic()
     return (llvmInternal == LLVMva_start ||
             llvmInternal == LLVMva_copy ||
             llvmInternal == LLVMva_end);
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////
+
+void emitABIReturnAsmStmt(IRAsmBlock* asmblock, Loc loc, FuncDeclaration* fdecl)
+{
+    Logger::println("emitABIReturnAsmStmt(%s)", fdecl->mangle());
+    LOG_SCOPE;
+
+    IRAsmStmt* as = new IRAsmStmt;
+
+    LLType* llretTy = DtoType(fdecl->type->nextOf());
+    asmblock->retty = llretTy;
+    asmblock->retn = 1;
+
+    // FIXME: This should probably be handled by the TargetABI somehow.
+    // It should be able to do this for a greater variety of types.
+
+    // x86
+    if (global.params.targetTriple.getArch() == llvm::Triple::x86)
+    {
+        LINK l = fdecl->linkage;
+        assert((l == LINKd || l == LINKc || l == LINKwindows) && "invalid linkage for asm implicit return");
+
+        Type* rt = fdecl->type->nextOf()->toBasetype();
+        if (rt->isintegral() || rt->ty == Tpointer || rt->ty == Tclass || rt->ty == Taarray)
+        {
+            if (rt->size() == 8) {
+                as->out_c = "=A,";
+            } else {
+                as->out_c = "={ax},";
+            }
+        }
+        else if (rt->isfloating())
+        {
+            if (rt->iscomplex()) {
+                if (fdecl->linkage == LINKd) {
+                    // extern(D) always returns on the FPU stack
+                    as->out_c = "={st},={st(1)},";
+                    asmblock->retn = 2;
+                } else if (rt->ty == Tcomplex32) {
+                    // extern(C) cfloat is return as i64
+                    as->out_c = "=A,";
+                    asmblock->retty = LLType::getInt64Ty(gIR->context());
+                } else {
+                    // cdouble and creal extern(C) are returned in pointer
+                    // don't add anything!
+                    asmblock->retty = LLType::getVoidTy(gIR->context());
+                    asmblock->retn = 0;
+                    return;
+                }
+            } else {
+                as->out_c = "={st},";
+            }
+        }
+        else if (rt->ty == Tarray || rt->ty == Tdelegate)
+        {
+            as->out_c = "={ax},={dx},";
+            asmblock->retn = 2;
+        #if 0
+            // this is to show how to allocate a temporary for the return value
+            // in case the appropriate multi register constraint isn't supported.
+            // this way abi return from inline asm can still be emulated.
+            // note that "$<<out0>>" etc in the asm will translate to the correct
+            // numbered output when the asm block in finalized
+
+            // generate asm
+            as->out_c = "=*m,=*m,";
+            LLValue* tmp = DtoRawAlloca(llretTy, 0, ".tmp_asm_ret");
+            as->out.push_back( tmp );
+            as->out.push_back( DtoGEPi(tmp, 0,1) );
+            as->code = "movd %eax, $<<out0>>" "\n\t" "mov %edx, $<<out1>>";
+
+            // fix asmblock
+            asmblock->retn = 0;
+            asmblock->retemu = true;
+            asmblock->asmBlock->abiret = tmp;
+
+            // add "ret" stmt at the end of the block
+            asmblock->s.push_back(as);
+
+            // done, we don't want anything pushed in the front of the block
+            return;
+        #endif
+        }
+        else
+        {
+            error(loc, "unimplemented return type '%s' for implicit abi return", rt->toChars());
+            fatal();
+        }
+    }
+
+    // x86_64
+    else if (global.params.targetTriple.getArch() == llvm::Triple::x86_64)
+    {
+        LINK l = fdecl->linkage;
+        /* TODO: Check if this works with extern(Windows), completely untested.
+* In particular, returning cdouble may not work with
+* extern(Windows) since according to X86CallingConv.td it
+* doesn't allow XMM1 to be used.
+* (So is extern(C), but that should be fine as the calling convention
+* is identical to that of extern(D))
+*/
+        assert((l == LINKd || l == LINKc || l == LINKwindows) && "invalid linkage for asm implicit return");
+
+        Type* rt = fdecl->type->nextOf()->toBasetype();
+        if (rt->isintegral() || rt->ty == Tpointer || rt->ty == Tclass || rt->ty == Taarray)
+        {
+            as->out_c = "={ax},";
+        }
+        else if (rt->isfloating())
+        {
+            if (rt == Type::tcomplex80) {
+                // On x87 stack, re=st, im=st(1)
+                as->out_c = "={st},={st(1)},";
+                asmblock->retn = 2;
+            } else if (rt == Type::tfloat80 || rt == Type::timaginary80) {
+                // On x87 stack
+                as->out_c = "={st},";
+            } else if (l != LINKd && rt == Type::tcomplex32) {
+                // LLVM and GCC disagree on how to return {float, float}.
+                // For compatibility, use the GCC/LLVM-GCC way for extern(C/Windows)
+                // extern(C) cfloat -> %xmm0 (extract two floats)
+                as->out_c = "={xmm0},";
+                asmblock->retty = LLType::getDoubleTy(gIR->context());
+            } else if (rt->iscomplex()) {
+                // cdouble and extern(D) cfloat -> re=%xmm0, im=%xmm1
+                as->out_c = "={xmm0},={xmm1},";
+                asmblock->retn = 2;
+            } else {
+                // Plain float/double/ifloat/idouble
+                as->out_c = "={xmm0},";
+            }
+        }
+        else if (rt->ty == Tarray || rt->ty == Tdelegate)
+        {
+            as->out_c = "={ax},={dx},";
+            asmblock->retn = 2;
+        }
+        else
+        {
+            error(loc, "unimplemented return type '%s' for implicit abi return", rt->toChars());
+            fatal();
+        }
+    }
+
+    // unsupported
+    else
+    {
+        error(loc, "this target (%s) does not implement inline asm falling off the end of the function", global.params.targetTriple.str().c_str());
+        fatal();
+    }
+
+    // return values always go in the front
+    asmblock->s.push_front(as);
 }
