+++ gen/functions.h
@@ -37,7 +37,6 @@ void DtoResolveFunction(FuncDeclaration* fdecl);
 void DtoDeclareFunction(FuncDeclaration* fdecl);
 void DtoDefineFunction(FuncDeclaration* fd);
 
-void DtoDefineNakedFunction(FuncDeclaration* fd);
 void emitABIReturnAsmStmt(IRAsmBlock* asmblock, Loc loc, FuncDeclaration* fdecl);
 
 DValue* DtoArgument(Parameter* fnarg, Expression* argexp);
