+++ gen/tollvm.cpp
@@ -344,10 +344,8 @@ LLGlobalValue::LinkageTypes DtoLinkage(Dsymbol* sym)
             return llvm::GlobalValue::ExternalLinkage;
 
         // Mark functions generated by an inlining semantic run as
-        // available_externally. Naked functions are turned into module-level
-        // inline asm and are thus declaration-only as far as the LLVM IR level
-        // is concerned.
-        if (fdecl->availableExternally && mustDefine && !fdecl->naked)
+        // available_externally.
+        if (fdecl->availableExternally && mustDefine)
             return llvm::GlobalValue::AvailableExternallyLinkage;
 
         // array operations are always template linkage
@@ -355,9 +353,8 @@ LLGlobalValue::LinkageTypes DtoLinkage(Dsymbol* sym)
             return templateLinkage;
 
         // template instances should have weak linkage
-        // but only if there's a body, and it's not naked
-        // otherwise we make it external
-        if (needsTemplateLinkage(fdecl) && fdecl->fbody && !fdecl->naked)
+        // but only if there's a body
+        if (needsTemplateLinkage(fdecl) && fdecl->fbody)
             return templateLinkage;
 
         // extern(C) functions are always external
